# docs/swagger.yaml
definitions:
  models.ErrorDetail:
    properties:
      code:
        type: string
      correlation_id:
        type: string
      message:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.ErrorDetail'
    type: object
...
  models.Response:
    properties:
      data: {}
    type: object
info:
  contact: {}
  description: KVM/QEMU VM management platform API
  title: VirtueStack API
  version: "1.0"
paths:
...
  /api/v1/provisioning/tasks/{id}:
    get:
      description: Returns provisioning task status by task ID.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - APIKeyAuth: []
      summary: Get task status
      tags:
      - Provisioning
...
securityDefinitions:
  APIKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"